Document in _.memoize why _.identity is a first-argument hasher
With a comment, document the nonobvious behavior of _.identity in this context.

I wrote this comment to accompany the following sentence from [the current docs for `_.memoize`](http://documentcloud.github.io/underscore/#memoize):

> The default **hashFunction** just uses the first argument to the memoized function as the key.

I looked in the code, but couldn’t see how that was the case. The code quoted in [Issue #137](https://github.com/jashkenas/underscore/issues/137) was no longer there. I wondered if that sentence in the documentation was out-of-date. It took me a while to notice that the use of `apply` over `call` changed `_.identity` from returning `arguments` to returning the first argument.