clarified wrap() documentation
Initial example was confusing to someone newer programers. New example highlights the ability of _.wrap to have a function execute code before and after a function call, without adding the extra step of having having a function redefine itself as a wrapped version of it's previous self. By assigning the function to a new variable name, the _.wrap function is more explicit.