_.max/_.min guard specs use deepEqual [fix #2037]
The specs for testing _.max and _.min now work correctly after switching
equal to deepEqual.

Passing _.max and _.min into _.map as an iteratee can be used to
retrieve the maximum and minimum numbers from an array of arrays. It
should be noted however that this functionaliy will not work as expected
if an array of objects is passed in.