_.matches only validates own properties[fix #1986]
I have changed _.matches to only match properties which are declared
directly on the object being passed to the predicate function generated
by _.matches.

I have accomplished this by testing the property on the object passed in
using the method hasOwnProperty method instead of the in operator.

I have included specs to demonstrate how this functionality works and
when it should return true or false.