Add a purpose section to README.md
To make technologyRepo more approachable to a larger number of users—from informed developers to curious passersby and new and interested visitors—I have added a new section to the technologyRepo README. The section is entitled What to Use technologyRepo for and When to Use It. As the name suggests, the new section, a brief paragraph, explains in plain and straightforward language what underscore.js is used for and when one should use it.

Other repositories may begin to use this new informative and catchy section soon, as my colleagues contact open-source technologies in an effort to make them more accessible to the tens of thousands of new users learning to code and thousands joining GitHub every day.