Allow users to customize _.iteratee
This fix was originally proposed by @jridgewell in #1965 but it was unable to
gain consensus since his approach added an additional layer of abstraction
to which @akre54 objected. I've refactored @jridgewell's approach to avoid that
abstraction.

This PR was motivated by a renewed interest in customizable shorthand syntaxes
such as the RegEx syntax proposed in #2475.