Fix debounce immediate
Fixes an issue with `_.debounce` when `{ immediate: true }`. When
called twice within `wait` ms, we cleared the `timeout`. Because it is
truthy on the second run, it was cleared but `timeout` remained
truthy (on any later run, `timeout` remains truthy). Because `timeout`
was cleared, the `later` function is never run to null out `timeout`.

Anyways, we should always be setting a fresh timeout function when
`{ immediate: true }`, to prevent further calls to the debounced
function until the `wait` ms after the last call.

Fixes #2478, supersedes #2479. Thanks for the bug report @hanzichi.