Remove no-extra-parens ESLint config rule
The no-extra-parens ESLint rule was causing the test suite to throw warnings. But the location of the warning did not look like it had any problematic parenths. This is likely a bug with ESLint. Till it is resolved, I think it makes sense to remove this flag to clean up the test suite logs. This will also not have any effect on the code anyway, as extra parentheses are ignored by most JS Engines