Remove test for edge case: _.reject(list, []);
I believe this test was added by accident by @megawac in #1582
(specifically ca404d40f1705ffc84c5cd8a8a95a04373df2c76)

Given our current documentation, the behavior of this edge case is not
defined.

This test currently passes because arrays are technically objects and
thus `_.iteratee` treats `[]` as a `_.matches` matchers. `_.matches`
casts `[]` to `{}` via `_.extendOwn({}, []);` and returns a function
witch always returns `true`.

The fact that this test exists in `_.reject`'s test and not in
`_.iteratee`'s leads me to believe that this is not actually intended
behavior, but merely an implementation detail which has been
accidentally enforced in our tests.

I'm currently reworking my pull request #2494 (deep property access) and
in a future where arrays are used for deep property access, `[]` will
return a function that always returns `undefined`, which seems like
a more reasonable behavior.

I thought it would be good to have this conversation separate from the
deep property matching pull request, but removing this test is
a prerequisite to implementing deep property access.