Don't cache Date.now
Certain test frameworks override Date.now to match their stubbed
versions of `setTimeout`, `setInterval`, etc.  While Underscore works
well with replaced versions of those function (because it gets the
current copy each time they are used), this isn't true of the `Date.
now` function, which is cached when Underscore is first initialized.

This change ensures that Underscore uses the current version of the
`Date.now` function each time it is called for, and therefore
improves compatibility with such test libraries.